address pull request comments
