+++ core/src/main/java/org/apache/accumulo/core/conf/Property.java
@@ -530,8 +530,8 @@ public enum Property {
           + " to force the changes to be written to disk, the metadata and root tables can be flushed"
           + " and possibly compacted. Legal values are: compact - which both flushes and compacts the"
           + " metadata; flush - which flushes only (compactions may be triggered if required); or none"),
-  GC_ENABLE_METRICS2("gc.enable.metrics2", "true", PropertyType.BOOLEAN,
-      "Enable detailed gc metrics reporting with hadoop metrics2."),
+  GC_METRICS_ENABLED("gc.enable.metrics", "true", PropertyType.BOOLEAN,
+      "Enable detailed gc metrics reporting with hadoop metrics."),
 
   // properties that are specific to the monitor server behavior
   MONITOR_PREFIX("monitor.", null, PropertyType.PREFIX,
